Disconnected node case

+++ src/power/execution/BlueprintNode.hpp
@@ -142,7 +142,7 @@ public:
 		return mData;
 	}
 	
-	virtual void set_data(std::optional<std::variant<Entity, std::string, int, float, bool>> data) {
+	virtual void set_data(std::optional<std::variant<Entity, std::string, long, float, bool>> data) {
 		mData = data;
 	}
 	
@@ -257,9 +257,9 @@ public:
 	DataCoreNode(NodeType type, UUID id, nanogui::Color color = nanogui::Color(255, 255, 255, 255)) : CoreNode(type, id, color) {
 		
 	}
-	virtual std::optional<std::variant<Entity, std::string, int, float, bool>> get_data() const = 0;
+	virtual std::optional<std::variant<Entity, std::string, long, float, bool>> get_data() const = 0;
 	
-	virtual void set_data(std::optional<std::variant<Entity, std::string, int, float, bool>> data) = 0;
+	virtual void set_data(std::optional<std::variant<Entity, std::string, long, float, bool>> data) = 0;
 };
