+++ src/power/execution/StringNode.hpp
@@ -13,9 +13,9 @@ class StringCoreNode : public DataCoreNode {
 public:
 	StringCoreNode(UUID id);
 
-	void set_data(std::optional<std::variant<Entity, std::string, int, float, bool>> data) override;
+	void set_data(std::optional<std::variant<Entity, std::string, long, float, bool>> data) override;
 
-	std::optional<std::variant<Entity, std::string, int, float, bool>> get_data() const override;
+	std::optional<std::variant<Entity, std::string, long, float, bool>> get_data() const override;
 	
 private:
 	CorePin& mOutput;
